fix: fix mqtt feature dependencies
mqtt feature required examples feature, which pulled olink feature,
so having mqtt feature result in having a code for olink, with all its
dependencies not required for mqtt.
This was fixed with:
Examples were split into examples_olink and examples_mqtt.
Examples are not required mqtt, and need to be explicite requested if
expected.
The olink is not required neither for mqtt nor for examples_mqtt.
This closes #86