chore: move cmake dependencies to top cmake
Clean up Catch2 dependencies it was already used with fetch content
in 2 CMakeLists files.
This commits moves fetching to toplevel cmake, and requires this package
in projects that uses it.
The Catch2 will be also used in generated olink tests for interface
olink adapters.
This is preparatory commit for adding generated olink tests #90.