feat(olink): add method tests
introduces test skeletons for method and their async versions.
It uses default values as, without mocks it is not easy to check on the
service or implementation side. The return value is set to default in
the generated implementation. In current setup this is as much as it can
be tested.
This is part of #90