feat: add test helpers
Introduces new rules feature: test_helpers.
It helps to prepare test structure - one with non default values, to
allow testing of properties and signals that use those types.
It uses qtTestValue filter which can fill simple type (or returns the
empty structure in case of array of structure type).
This may be useful for testing generated interfaces: the implementation,
the olink, mqtt or any other adapter.
This is preparatory change for implementing tests for olink adapters #90