feat(olink): test olink adapters set property
Add testing set property for generated adapters.
The tests are setting up a client and a service adapter which exposes
the local default implementation - this is why stubs are added as
requirement for the olink feature.
The client and server are connected
without network (client writes to servers read function and vice versa).
This is part of #90.