feat(mqtt): add generated tests project
introduces test project for mqtt - the tests will be added with next
commit.
This project is set to run only on linux machine by default.
To run on other os use an option for it in top level CMakeLists.
This is due to the CI which has mqtt broker set and run only for linux.
This is preparatory commit for #89 task.