fix(mqtt): use better names for mqtt callbacks
this is preparatory commit for handling change of mqtt subscription
topic.
The name subscribeCallback name is a bit misleading as the variable
of this type will be holding handler for simple handling (just passing a
message arguments) for incoming message for given topic.
In the next commits,the subscription function will be extended with
another callback to handle finishing of subscribing process, and the
name change will help understande the difference between them.
Also this commit improves consistency:
the callback types start with upper letter
the callback types are used in all the places they should be used.

This is part of #89