fix(mqtt): handle on subscribe topic
The mqtt protocol and its implementation in qt classes inform about the
subscription status, if it is already finished and with what result.
This information is used by the apigear adapters, it is required to
establish when the generated adapters are ready to receive messages -
the subscription for topic must be done. Otherwise, sending a
request-response messages may fail in a way, the object won't subscribe
for the response before it arrive (it is not very likely in the network,
but it is easy to reproduce in testing).
This commit introduces the functions in general client adapters for
apigear mqtt client and service, and forces the generated adapters for
each interface to pass their handlers.
The handlers are implemented for the generated interface adapters.
This is partial commit, next change adds using the handlers in the
client and service.
This is part of #89